refactor: separate User and UserDetail with Mapstruct
