+++ src/main/java/com/tech_symfony/movie_booking/api/user/CustomUserDetailService.java
@@ -1,6 +1,7 @@
 package com.tech_symfony.movie_booking.api.user;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -8,7 +9,6 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-
 public class CustomUserDetailService implements UserDetailsService {
 
 	private final UserRepository userRepository;
@@ -17,7 +17,6 @@ public class CustomUserDetailService implements UserDetailsService {
 	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
 		User user = userRepository.findByEmail(username)
 			.orElseThrow(() -> new UsernameNotFoundException("User not found"));
-//		return user;
-		return user;
+		return UserMapper.INSTANCE.userToCustomerUserDetail(user);
 	}
 }
