+++ src/main/java/com/tech_symfony/movie_booking/api/user/UserMapper.java
@@ -1,0 +1,15 @@
+package com.tech_symfony.movie_booking.api.user;
+
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.factory.Mappers;
+
+@Mapper
+public interface UserMapper {
+	UserMapper INSTANCE = Mappers.getMapper(UserMapper.class);
+
+	@Mapping(target = "principle", source = "email")
+	@Mapping(target = "credential", source = "password")
+	@Mapping(target = "authorities", ignore = true)
+	CustomUserDetail userToCustomerUserDetail(User user);
+}
